apps/api/pyproject.toml
[project]
name = "personal-finance-api"
version = "0.1.0"
description = "Personal Finance API backend"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Developer" }]

dependencies = [
    # Core
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.7.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    # Supporting
    "structlog>=24.4.0",
    "cryptography>=43.0.0",
    "typeid-python>=0.3.0",
    "returns>=0.23.0",
    # Authentication
    "PyJWT>=2.11.0",
    "pwdlib[argon2]>=0.3.0",
    "itsdangerous>=2.2.0",
    "python-multipart>=0.0.9",
    # Domain Events / Job Queue
    "procrastinate>=3.5.0",
    "psycopg[binary]>=3.2.0",
    # Email
    "Jinja2>=3.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
